Refactor code into modules
